+++ geoservices/src/main/java/com/jeltechnologies/geoservices/datasources/house/GeoCoordinates.java
@@ -6,7 +6,7 @@ import com.jeltechnologies.geoservices.datasources.Location;
 public class GeoCoordinates extends Location {
     private final int id;
 
-    public GeoCoordinates(int id, float lat, float lon) {
+    public GeoCoordinates(int id, double lat, double lon) {
 	super(new Coordinates(lat, lon));
 	this.id = id;
     }
